requirements.txt
pkg-resources==0.0.0
python-decouple==3.4
